fix: Rebased location and moved files into correct folder
