+++ app/code/Kai/Googlefeed/etc/adminhtml/routes.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
-    <router id="admin">
-        <route id="kaigooglefeed" frontName="kaigooglefeed">
-            <module name="Kai_Googlefeed"/>
-        </route>
-    </router>
+	<router id="admin">
+		<route id="kaigooglefeed" frontName="kaigooglefeed">
+			<module name="Kai_Googlefeed"/>
+		</route>
+	</router>
 </config>
