+++ app/code/Kai/Googlefeed/etc/module.xml
@@ -1,4 +1,4 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Kai_Googlefeed" setup_version="1.0.0">
-    </module>
+	<module name="Kai_Googlefeed" setup_version="1.0.0">
+	</module>
 </config>
