+++ app/code/Kai/Googlefeed/registration.php
@@ -1,0 +1,6 @@
+<?php
+\Magento\Framework\Component\ComponentRegistrar::register(
+	\Magento\Framework\Component\ComponentRegistrar::MODULE,
+	'Kai_Googlefeed',
+	__DIR__
+);
